A bit more formatting
Organized into one file for each part, Thinking about the jupyter notebook requirement. As far as I know it doesn't have a good feature control system, so maybe if we could find some way to download and edit jupyter files that would allow it to mesh well with git. I'll look into it probably soon but if anyone knows anything about it feel free to mention.