Cleaning up
Re-merged Part1 and Part1Testing, added a bunch of comments and a test case for the function passing business. I looked up list comprehensions and realize I must have been the victim of some kind of scaremongering at some point because they are entirely trustworthy and order-preserving on any kind of ordered class. Will be adding more comments and the like soon tonight.